Improve cli structure
